Chore: Restructure typescript tests into test directory
